refactored validatorUtils
